+++ sphinx/source/changelog.rst
@@ -14,6 +14,8 @@ Changelog
   issue 2 emulation)
 * Added support to :ref:`tap2sna.py <tap2sna-conf>` for reading configuration
   from `skoolkit.ini`
+* Added the ``--ini`` option to :ref:`tap2sna.py` (for setting the value of a
+  configuration parameter)
 * :ref:`snapinfo.py` now shows the value of the T-states counter in SZX and Z80
   snapshots
 * Fixed the bug that prevents :ref:`snapinfo.py` from displaying the value of a